requirements.txt
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.42
pyinstaller==6.14.2
pyinstaller-hooks-contrib==2025.8
pystray==0.19.5
pillow==11.3.0
